# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Listener
  description: Golang HTTP service listening for Callback and Timeout URLs
  termsOfService: https://github.com/wondenge/mpesa-listener/blob/master/LICENSE
  contact:
    name: William Ondenge
    email: ondengew@gmail.com
    url: https://www.ondenge.me
  license:
    name: Apache License
    url: https://github.com/wondenge/mpesa-listener/blob/master/LICENSE
  version: "1.0"
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /mpesa/accountbalance/v1/result:
    post:
      tags:
      - mpesa
      summary: AccountBalanceResult mpesa
      description: Account Balance Result URL
      operationId: mpesa#AccountBalanceResult
      parameters:
      - name: AccountBalanceResultRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaAccountBalanceResultRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/accountbalance/v1/timeout:
    post:
      tags:
      - mpesa
      summary: AccountBalanceTimeout mpesa
      description: Account Balance Queue TimeOut URL
      operationId: mpesa#AccountBalanceTimeout
      parameters:
      - name: AccountBalanceTimeoutRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaAccountBalanceTimeoutRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/b2c/v1/result:
    post:
      tags:
      - mpesa
      summary: B2CResult mpesa
      description: B2C Result URL
      operationId: mpesa#B2CResult
      parameters:
      - name: B2CResultRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaB2CResultRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/b2c/v1/timeout:
    post:
      tags:
      - mpesa
      summary: B2CTimeout mpesa
      description: B2C Queue TimeOut URL
      operationId: mpesa#B2CTimeout
      parameters:
      - name: B2CTimeoutRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaB2CTimeoutRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/c2b/v1/confirmation:
    post:
      tags:
      - mpesa
      summary: C2BConfirmation mpesa
      description: C2B Confirmation URL
      operationId: mpesa#C2BConfirmation
      parameters:
      - name: C2BConfirmationRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaC2BConfirmationRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/c2b/v1/validation:
    post:
      tags:
      - mpesa
      summary: C2BValidation mpesa
      description: C2B Validation URL
      operationId: mpesa#C2BValidation
      parameters:
      - name: C2BValidationRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaC2BValidationRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/reversal/v1/result:
    post:
      tags:
      - mpesa
      summary: ReversalResult mpesa
      description: Reversal Result URL
      operationId: mpesa#ReversalResult
      parameters:
      - name: ReversalResultRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaReversalResultRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/reversal/v1/timeout:
    post:
      tags:
      - mpesa
      summary: ReversalTimeout mpesa
      description: Reversal Queue TimeOut URL
      operationId: mpesa#ReversalTimeout
      parameters:
      - name: ReversalTimeoutRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaReversalTimeoutRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/transactionstatus/v1/result:
    post:
      tags:
      - mpesa
      summary: TransactionStatusResult mpesa
      description: Transaction Status Result URL
      operationId: mpesa#TransactionStatusResult
      parameters:
      - name: TransactionStatusResultRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaTransactionStatusResultRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /mpesa/transactionstatus/v1/timeout:
    post:
      tags:
      - mpesa
      summary: TransactionStatusTimeout mpesa
      description: Transaction Status Queue TimeOut URL
      operationId: mpesa#TransactionStatusTimeout
      parameters:
      - name: TransactionStatusTimeoutRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/MpesaTransactionStatusTimeoutRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
...
definitions:
  MpesaAccountBalanceResultRequestBody:
    title: MpesaAccountBalanceResultRequestBody
    type: object
    properties:
      ConversationID:
        type: string
        example: AG_20170728_0000589b6252f7f25488
      OriginatorConversationID:
        type: string
        example: 19464-802673-1
      ReferenceData:
        type: object
        properties:
          ReferenceItem:
            type: object
            example:
              QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
            additionalProperties: true
        example:
          ReferenceItem:
            QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
      ResultCode:
        type: integer
        example: 0
        format: int64
      ResultDesc:
        type: string
        example: The service request has b een accepted successfully.
      ResultParameters:
        type: object
        properties:
          ResultParameter:
            type: object
            properties:
              AccountBalance:
                type: object
                example:
                  AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
                additionalProperties: true
              BOCompletedTime:
                type: object
                example:
                  BOCompletedTime: "20170728095642"
                additionalProperties: true
            example:
              AccountBalance:
                AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
              BOCompletedTime:
                BOCompletedTime: "20170728095642"
        example:
          ResultParameter:
            AccountBalance:
              AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
            BOCompletedTime:
              BOCompletedTime: "20170728095642"
      ResultType:
        type: integer
        example: 0
        format: int64
      TransactionID:
        type: string
        example: LGS0000000
    example:
      ConversationID: AG_20170728_0000589b6252f7f25488
      OriginatorConversationID: 19464-802673-1
      ReferenceData:
        ReferenceItem:
          QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
      ResultCode: 0
      ResultDesc: The service request has b een accepted successfully.
      ResultParameters:
        ResultParameter:
          AccountBalance:
            AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
          BOCompletedTime:
            BOCompletedTime: "20170728095642"
      ResultType: 0
      TransactionID: LGS0000000
  MpesaAccountBalanceTimeoutRequestBody:
    title: MpesaAccountBalanceTimeoutRequestBody
    type: object
    properties:
      ConversationID:
        type: string
        example: AG_20170728_0000589b6252f7f25488
      OriginatorConversationID:
        type: string
        example: 19464-802673-1
      ReferenceData:
        type: object
        properties:
          ReferenceItem:
            type: object
            example:
              QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
            additionalProperties: true
        example:
          ReferenceItem:
            QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
      ResultCode:
        type: integer
        example: 0
        format: int64
      ResultDesc:
        type: string
        example: The service request has b een accepted successfully.
      ResultParameters:
        type: object
        properties:
          ResultParameter:
            type: object
            properties:
              AccountBalance:
                type: object
                example:
                  AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
                additionalProperties: true
              BOCompletedTime:
                type: object
                example:
                  BOCompletedTime: "20170728095642"
                additionalProperties: true
            example:
              AccountBalance:
                AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
              BOCompletedTime:
                BOCompletedTime: "20170728095642"
        example:
          ResultParameter:
            AccountBalance:
              AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
            BOCompletedTime:
              BOCompletedTime: "20170728095642"
      ResultType:
        type: integer
        example: 0
        format: int64
      TransactionID:
        type: string
        example: LGS0000000
    example:
      ConversationID: AG_20170728_0000589b6252f7f25488
      OriginatorConversationID: 19464-802673-1
      ReferenceData:
        ReferenceItem:
          QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/abresults/v1/submit
      ResultCode: 0
      ResultDesc: The service request has b een accepted successfully.
      ResultParameters:
        ResultParameter:
          AccountBalance:
            AccountBalance: Working Account|KES|46713.00|46713.00|0.00|0.00&Float Account|KES|0.00|0.00|0.00|0.00&Utility Account|KES|49217.00|49217.00|0.00|0.00&Charges Paid Account|KES|-220.00|-220.00|0.00|0.00&Organization Settlement Account|KES|0.00|0.00|0.00|0.00
          BOCompletedTime:
            BOCompletedTime: "20170728095642"
      ResultType: 0
      TransactionID: LGS0000000
  MpesaB2CResultRequestBody:
    title: MpesaB2CResultRequestBody
    type: object
    properties:
      ReferenceData:
        type: object
        properties:
          ReferenceItem:
            type: object
            example:
              QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/b2cresults/v1/submit
            additionalProperties: true
        example:
          ReferenceItem:
            QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/b2cresults/v1/submit
      Result:
        type: object
        properties:
          ConversationId:
            type: string
            description: Global unique identifier for the transaction request.
            example: 236543-276372-2
          OriginatorConversationId:
            type: string
            description: Global unique identifier for the transaction request.
            example: AG_2376487236_126732989KJHJKH
          ResultCode:
            type: integer
            description: Numeric status code that indicates the status of the transaction processing
            example: 0
            format: int64
          ResultDesc:
            type: string
            description: Message from the API that gives the status of the request
            example: Service request is has bee accepted successfully
            enum:
            - Service request is has bee accepted successfully
            - Initiator information is invalid
          ResultType:
            type: integer
            description: Status code that indicates whether the transaction was already sent
            default: 0
            example: 5788603017554224693
            format: int64
          TransactionID:
            type: string
            description: Unique M-PESA transaction ID for every payment request.
            example: LHG31AA5TX
        example:
          ConversationId: 236543-276372-2
          OriginatorConversationId: AG_2376487236_126732989KJHJKH
          ResultCode: 0
          ResultDesc: Initiator information is invalid
          ResultType: 6495929187470417453
          TransactionID: LHG31AA5TX
      ResultParameters:
        type: object
        properties:
          ResultParameter:
            type: object
            properties:
              B2CChargesPaidAccountAvailableFunds:
                type: object
                description: Available balance of the Charges Paid account under the B2C shortcode used in the transaction.
                example:
                  B2CChargesPaidAccountAvailableFunds: 0
                additionalProperties: true
              B2CRecipientIsRegisteredCustomer:
                type: object
                description: This is a key that indicates whether the customer is a M-PESA registered customer or not
                example:
                  ReceiverPartyPublicName: "Y"
                additionalProperties: true
              B2CUtilityAccountAvailableFunds:
                type: object
                description: Available balance of the Utility account under the B2C shortcode used in the transaction.
                example:
                  B2CUtilityAccountAvailableFunds: 133568
                additionalProperties: true
              B2CWorkingAccountAvailableFunds:
                type: object
                description: Available balance of the Working account under the B2C shortcode used in the transaction.
                example:
                  B2CWorkingAccountAvailableFunds: 150000
                additionalProperties: true
              ReceiverPartyPublicName:
                type: object
                description: This is the name and phone number of the customer who received the payment.
                example:
                  ReceiverPartyPublicName: 254708374149 - John Doe
                additionalProperties: true
              TransactionAmount:
                type: object
                description: This is the amount that was transacted.
                example:
                  TransactionAmount: 8000
                additionalProperties: true
              TransactionCompletedDateTime:
                type: object
                description: This is the date and time that the transaction completed M-PESA.
                example:
                  TransactionCompletedDateTime: 17.07.2017 10:54:57
                additionalProperties: true
              TransactionReceipt:
                type: object
                description: This is a unique M-PESA transaction ID for every payment request.
                example:
                  TransactionReceipt: LGH3197RIB
                additionalProperties: true
            example:
              B2CChargesPaidAccountAvailableFunds:
                B2CChargesPaidAccountAvailableFunds: 0
              B2CRecipientIsRegisteredCustomer:
                ReceiverPartyPublicName: "Y"
              B2CUtilityAccountAvailableFunds:
                B2CUtilityAccountAvailableFunds: 133568
              B2CWorkingAccountAvailableFunds:
                B2CWorkingAccountAvailableFunds: 150000
              ReceiverPartyPublicName:
                ReceiverPartyPublicName: 254708374149 - John Doe
              TransactionAmount:
                TransactionAmount: 8000
              TransactionCompletedDateTime:
                TransactionCompletedDateTime: 17.07.2017 10:54:57
              TransactionReceipt:
                TransactionReceipt: LGH3197RIB
        example:
          ResultParameter:
            B2CChargesPaidAccountAvailableFunds:
              B2CChargesPaidAccountAvailableFunds: 0
            B2CRecipientIsRegisteredCustomer:
              ReceiverPartyPublicName: "Y"
            B2CUtilityAccountAvailableFunds:
              B2CUtilityAccountAvailableFunds: 133568
            B2CWorkingAccountAvailableFunds:
              B2CWorkingAccountAvailableFunds: 150000
            ReceiverPartyPublicName:
              ReceiverPartyPublicName: 254708374149 - John Doe
            TransactionAmount:
              TransactionAmount: 8000
            TransactionCompletedDateTime:
              TransactionCompletedDateTime: 17.07.2017 10:54:57
            TransactionReceipt:
              TransactionReceipt: LGH3197RIB
    example:
      ReferenceData:
        ReferenceItem:
          QueueTimeoutURL: https://internalsandbox.safaricom.co.ke/mpesa/b2cresults/v1/submit
      Result:
        ConversationId: 236543-276372-2
        OriginatorConversationId: AG_2376487236_126732989KJHJKH
        ResultCode: 0
        ResultDesc: Initiator information is invalid
        ResultType: 7139382298279117317
        TransactionID: LHG31AA5TX
      ResultParameters:
        ResultParameter:
          B2CChargesPaidAccountAvailableFunds:
            B2CChargesPaidAccountAvailableFunds: 0
          B2CRecipientIsRegisteredCustomer:
            ReceiverPartyPublicName: "Y"
          B2CUtilityAccountAvailableFunds:
            B2CUtilityAccountAvailableFunds: 133568
          B2CWorkingAccountAvailableFunds:
            B2CWorkingAccountAvailableFunds: 150000
          ReceiverPartyPublicName:
            ReceiverPartyPublicName: 254708374149 - John Doe
          TransactionAmount:
            TransactionAmount: 8000
          TransactionCompletedDateTime:
            TransactionCompletedDateTime: 17.07.2017 10:54:57
          TransactionReceipt:
            TransactionReceipt: LGH3197RIB
...
  MpesaC2BConfirmationRequestBody:
    title: MpesaC2BConfirmationRequestBody
    type: object
    properties:
      BillRefNumber:
        type: string
        description: Account number for which the customer is making the payment
        example: h6h
        maxLength: 20
      BusinessShortCode:
        type: integer
        description: Organizations shortcode (Paybill or Buygoods)
        example: 654321
        format: int64
      FirstName:
        type: string
        description: First Name of the customer making the payment.
        example: John
      InvoiceNumber:
        type: string
        example: Velit at at non.
      LastName:
        type: string
        description: Last Name of the customer making payment.
        example: Jane
      MSISDN:
        type: integer
        description: Mobile number of the customer making the payment.
        example: 6289845049800592044
        format: int64
      MiddleName:
        type: string
        description: Middle Name of the customer making the payment.
        example: Doe
      OrgAccountBalance:
        type: integer
        description: Current utility account balance of the payment receiving organization shortcode
        example: 30671
        format: int64
      ThirdPartyTransID:
        type: string
        example: Eius id corporis ut consequatur nobis qui.
      TransAmount:
        type: integer
        description: The Amount transacted normally a numeric value
        example: 100
        format: int64
      TransID:
        type: string
        description: Unique M-Pesa transaction ID for every payment request
        example: LHG31AA5TX
      TransTime:
        type: string
        description: Timestamp of the transaction
        example: "20180713154301"
        format: date-time
      TransactionType:
        type: string
        description: The transaction type specified during the payment request.
        example: Pay Bill
        enum:
        - Buy Goods
        - Pay Bill
    example:
      BillRefNumber: sw9
      BusinessShortCode: 654321
      FirstName: John
      InvoiceNumber: Et temporibus nisi nobis corrupti tempora.
      LastName: Jane
      MSISDN: 9179890662714884691
      MiddleName: Doe
      OrgAccountBalance: 30671
      ThirdPartyTransID: Atque culpa ducimus dolorem et minus aliquid.
      TransAmount: 100
      TransID: LHG31AA5TX
      TransTime: "20180713154301"
      TransactionType: Buy Goods
...